+++ server/migration/sqlite/1790200000000-AddBookOverrideRules.ts
@@ -4,14 +4,18 @@ export class AddBookOverrideRules1790200000000 implements MigrationInterface {
   name = 'AddBookOverrideRules1790200000000';
 
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(
-      `ALTER TABLE "override_rule" ADD "readarrServiceId" integer`
-    );
+    if (!(await queryRunner.hasColumn('override_rule', 'readarrServiceId'))) {
+      await queryRunner.query(
+        `ALTER TABLE "override_rule" ADD "readarrServiceId" integer`
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(
-      `ALTER TABLE "override_rule" DROP COLUMN "readarrServiceId"`
-    );
+    if (await queryRunner.hasColumn('override_rule', 'readarrServiceId')) {
+      await queryRunner.query(
+        `ALTER TABLE "override_rule" DROP COLUMN "readarrServiceId"`
+      );
+    }
   }
 }
